modularizing javascript to use ajax request from module
