+++ monopyly/static/js/autocomplete_card.js
@@ -6,6 +6,9 @@
  * being input using an AJAX request. 
  */
 
+import { executeAjaxRequest } from './modules/ajax.js';
+
+
 (function() {
 
 	// Identify all input elements in the form
@@ -41,24 +44,16 @@
 			'field': field,
 		};
 		// Return a set of autocomplete suggestions from the database
-		$.ajax({
-			url: AUTOCOMPLETE_ENDPOINT,
-			type: 'POST',
-			data: JSON.stringify(rawData),
-			contentType: 'application/json; charset=UTF-8',
-			success: function(response) {
-				// Define variables for the user input and its database matches
-				const userInput = $(inputElement).val().toLowerCase();
-				// Refresh the set of matches and update the menu
-				refreshMatches(userInput, response);
-				updateAutocomplete(inputElement, userInput);
-				// Bind keyup to also perform update, arrow keys to navigate
-				bindUpdate(inputElement, response);
-			},
-			error: function(xhr) {
-				console.log('There was an error in the Ajax request.');
-			}
-		});
+		function autocompleteAction(response) {
+			// Define variables for the user input and its database matches
+			const userInput = $(inputElement).val().toLowerCase();
+			// Refresh the set of matches and update the menu
+			refreshMatches(userInput, response);
+			updateAutocomplete(inputElement, userInput);
+			// Bind keyup to also perform update, arrow keys to navigate
+			bindUpdate(inputElement, response);
+		}
+		executeAjaxRequest(AUTOCOMPLETE_ENDPOINT, rawData, autocompleteAction);
 	}
 	
 	function unbindUpdate(inputElement) {
