+++ monopyly/static/js/modules/update_display_ajax.js
@@ -8,23 +8,7 @@
  * resuls of the request.
  */
 
-function executeAjaxRequest(endpoint, rawData, action = function(){} ) {
-
-	// Execute the action using the response of the AJAX request
-	$.ajax({
-		url: endpoint,
-		type: 'POST',
-		data: JSON.stringify(rawData),
-		contentType: 'application/json; charset=UTF-8',
-		success: function(response) {
-			action(response);
-		},
-		error: function(xhr) {
-			console.log('There was an error in the Ajax request.');
-		}
-	});
-
-}
+import { executeAjaxRequest } from './ajax.js';
 
 
 function replaceDisplayContentsAjaxRequest(endpoint, rawData, $display) {
@@ -39,4 +23,4 @@ function replaceDisplayContentsAjaxRequest(endpoint, rawData, $display) {
 }
 
 
-export { replaceDisplayContentsAjaxRequest, executeAjaxRequest };
+export { replaceDisplayContentsAjaxRequest };
