+++ monopyly/templates/credit/card_form_page.html
@@ -13,7 +13,7 @@
 	<script>
 		const AUTOCOMPLETE_ENDPOINT = "{{ url_for('credit.suggest_card_autocomplete') }}";
 	</script>
-	<script src="{{ url_for('static', filename='js/autocomplete_card.js') }}">
+	<script type="module" src="{{ url_for('static', filename='js/autocomplete_card.js') }}">
 	</script>
 
 {% endblock %}
