+++ monopyly/templates/credit/transaction_form/transaction_form_page.html
@@ -16,11 +16,11 @@
 	</script>
 	<script type="module" src="{{ url_for('static', filename='js/add_subtransaction.js') }}">
 	</script>
-	<script src="{{ url_for('static', filename='js/autocomplete_transaction.js') }}">
+	<script type="module" src="{{ url_for('static', filename='js/autocomplete_transaction.js') }}">
 	</script>
-	<script src="{{ url_for('static', filename='js/infer_card.js') }}">
+	<script type="module" src="{{ url_for('static', filename='js/infer_card.js') }}">
 	</script>
-	<script src="{{ url_for('static', filename='js/infer_statement.js') }}">
+	<script type="module" src="{{ url_for('static', filename='js/infer_statement.js') }}">
 	</script>
 
 {% endblock %}
